requirements.txt
flights>=0.7.0
python-telegram-bot[job-queue]>=21.0
aiosqlite>=0.20
python-dotenv>=1.0
